requirements.txt
rich==12.6.0
joblib==1.2.0
tqdm==4.64.1
numpy==1.23.4
pandas==1.5.1
websocket-client==1.4.2
pydantic==1.10.2
jsonlines==3.1.0
rel==0.4.7
pydantic-yaml==0.8.1
loguru==0.6.0
